fix: rename: teamcreate -> agendaCreate
